[region-isolation] Emit the correct error for closures that capture actor self.
rdar://122501400
(cherry picked from commit a9c163f8e2163365552317d6ab62e5e238c715d7)
